refactor(ast_tools): move logger into separate module (#6952)
Reduce the amount of code in `main.rs` by moving logger functionality to a separate module.
